\name{full_data}
\alias{full_data}
\docType{data}
\title{
Example data to illustrate functions
}
\description{
Simulated nonsmooth data to illustrate functions
}
\usage{data("full_data")}
\format{
  A data frame with 10100 observations on the following 5 variables.
  \describe{
    \item{\code{id}}{a unique person ID}
    \item{\code{a}}{treatment group, 0 or 1}
      \item{\code{tt}}{time}
     \item{\code{x}}{surrogate marker value}
       \item{\code{y}}{primary outcome}
  }
}

\keyword{datasets}
